+++ src/main/java/com/ll/gooHaeYu/domain/member/member/controller/MemberController.java
@@ -7,11 +7,14 @@ import com.ll.gooHaeYu.domain.jobPost.comment.service.CommentService;
 import com.ll.gooHaeYu.domain.jobPost.jobPost.dto.JobPostDto;
 import com.ll.gooHaeYu.domain.jobPost.jobPost.service.JobPostService;
 import com.ll.gooHaeYu.domain.member.member.dto.*;
+import com.ll.gooHaeYu.domain.member.member.service.AuthenticationService;
 import com.ll.gooHaeYu.domain.member.member.service.MemberService;
 import com.ll.gooHaeYu.global.rsData.RsData;
 import com.ll.gooHaeYu.global.security.MemberDetails;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -31,6 +34,7 @@ public class MemberController {
     private final JobPostService jobPostService;
     private final ApplicationService applicationService;
     private final CommentService commentService;
+    private final AuthenticationService authenticationService;
 
     @PostMapping("/join")
     @Operation(summary = "회원가입")
@@ -39,17 +43,18 @@ public class MemberController {
         return RsData.of("201", "CREATE", URI.create("/api/member/join" + userId));
     }
 
-    @PostMapping("/login")
-    @Operation(summary = "로그인, accessToken 쿠키 생성됨")
-    public RsData<String> login(@RequestBody @Valid LoginForm form) {
-
-        return RsData.of(memberService.login(form));
+    @PostMapping ("/login")
+    @Operation(summary = "로그인, accessToken, refreshToken 쿠키 생성됨")
+    public ResponseEntity<Void> login(@RequestBody @Valid LoginForm form, HttpServletRequest request, HttpServletResponse response) {
+        memberService.login(form);
+        authenticationService.authenticateAndSetTokens(form.getUsername(), request, response);
+        return ResponseEntity.ok().build();
     }
 
     @GetMapping
     @Operation(summary = "내 정보 조회")
     public RsData<MemberDto> detailMember(@AuthenticationPrincipal MemberDetails memberDetails) {
-        return  RsData.of(memberService.findByUsername(memberDetails.getUsername()));
+        return RsData.of(memberService.findByUsername(memberDetails.getUsername()));
     }
 
     @PutMapping
@@ -64,7 +69,7 @@ public class MemberController {
     @PutMapping("/social")
     @Operation(summary = "최초 소셜로그인 - 필수 회원정보 입력")
     public RsData<MemberDto> updateSocialMember(@AuthenticationPrincipal MemberDetails memberDetails,
-                                             @Valid @RequestBody SocialProfileForm form) {
+                                                @Valid @RequestBody SocialProfileForm form) {
         MemberDto updatedMember = memberService.updateSocialMemberProfile(memberDetails.getUsername(), form);
 
         return RsData.of(updatedMember);
@@ -73,20 +78,20 @@ public class MemberController {
     @GetMapping("/myposts")
     @Operation(summary = "내 공고 조회")
     public RsData<List<JobPostDto>> detailMyPosts(@AuthenticationPrincipal MemberDetails memberDetails) {
-        return  RsData.of(jobPostService.findByUsername(memberDetails.getUsername()));
+        return RsData.of(jobPostService.findByUsername(memberDetails.getUsername()));
     }
 
 
     @GetMapping("/myapplications")
     @Operation(summary = "내 지원서 조회")
     public RsData<List<ApplicationDto>> detailMyApplications(@AuthenticationPrincipal MemberDetails memberDetails) {
-        return  RsData.of(applicationService.findByUsername(memberDetails.getUsername()));
+        return RsData.of(applicationService.findByUsername(memberDetails.getUsername()));
     }
 
     @GetMapping("/mycomments")
     @Operation(summary = "내 댓글 조회")
     public RsData<List<CommentDto>> detailMyComments(@AuthenticationPrincipal MemberDetails memberDetails) {
-        return  RsData.of(commentService.findByUsername(memberDetails.getUsername()));
+        return RsData.of(commentService.findByUsername(memberDetails.getUsername()));
     }
 
     @GetMapping("/myinterest")
