+++ src/main/java/com/ll/gooHaeYu/domain/member/member/service/AuthenticationService.java
@@ -1,27 +1,22 @@
-package com.ll.gooHaeYu.global.security.OAuth;
+package com.ll.gooHaeYu.domain.member.member.service;
 
 import com.ll.gooHaeYu.domain.member.member.entity.Member;
 import com.ll.gooHaeYu.domain.member.member.entity.RefreshToken;
 import com.ll.gooHaeYu.domain.member.member.repository.RefreshTokenRepository;
-import com.ll.gooHaeYu.domain.member.member.service.MemberService;
 import com.ll.gooHaeYu.global.config.AppConfig;
 import com.ll.gooHaeYu.global.security.JwtTokenProvider;
 import com.ll.gooHaeYu.standard.base.util.CookieUtil;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
-import org.springframework.stereotype.Component;
-import org.springframework.web.util.UriComponentsBuilder;
+import org.springframework.stereotype.Service;
 
-import java.io.IOException;
 import java.time.Duration;
 
+@Service
 @RequiredArgsConstructor
-@Component
-public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler {
+public class AuthenticationService {
+
     public static final String REFRESH_TOKEN_COOKIE_NAME = "refresh_token";
     public static final String ACCESS_TOKEN_COOKIE_NAME = "access_token";
     public static final Duration REFRESH_TOKEN_DURATION = Duration.ofDays(1);
@@ -30,13 +25,10 @@ public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler 
 
     private final JwtTokenProvider jwtTokenProvider;
     private final RefreshTokenRepository refreshTokenRepository;
-    private final OAuth2AuthorizationRequestBasedOnCookieRepository authorizationRequestRepository;
     private final MemberService memberService;
 
-    @Override
-    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
-        OAuth2User oAuth2User = (OAuth2User) authentication.getPrincipal();
-        Member member = memberService.getMember((String) oAuth2User.getAttributes().get("email"));
+    public void authenticateAndSetTokens(String username, HttpServletRequest request, HttpServletResponse response) {
+        Member member = memberService.getMember(username);
 
         // 리프레쉬 토큰
         String refreshToken = jwtTokenProvider.generateToken(member, REFRESH_TOKEN_DURATION);
@@ -46,11 +38,7 @@ public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler 
         // 액세스 토큰
         String accessToken = jwtTokenProvider.generateToken(member, ACCESS_TOKEN_DURATION);
         addTokenToCookie(request, response, ACCESS_TOKEN_COOKIE_NAME, accessToken, ACCESS_TOKEN_DURATION);
-
-        clearAuthenticationAttributes(request, response);
-        getRedirectStrategy().sendRedirect(request, response, REDIRECT_PATH);
     }
-
     // 리프레쉬 토큰 DB 저장
     private void saveRefreshToken(Long userId, String newRefreshToken) {
         RefreshToken refreshToken = refreshTokenRepository.findByUserId(userId)
@@ -65,11 +53,5 @@ public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler 
         int cookieMaxAge = (int) duration.toSeconds();
         CookieUtil.deleteCookie(request, response, cookieName);
         CookieUtil.addCookie(response, cookieName, token, cookieMaxAge);
-    }
-
-    //인증 과정 중 생긴 요청 속성 제거
-    private void clearAuthenticationAttributes(HttpServletRequest request, HttpServletResponse response) {
-        super.clearAuthenticationAttributes(request);
-        authorizationRequestRepository.removeAuthorizationRequestCookies(request, response);
     }
 }
